fixed spinner
when we refactored the request chain we unfortunately lost an easy way to wait
for the chain to end. So the spinner would appear and disappear immediately
while the request chain did its work asynchronously.

Added `await` and `async function` to everything in the chain so we can get this
behaviour back.

Perhaps there's a cleaner solution.