follow PML spec in repo
Parser and lexer have been changed to adhere to the PML sepcification given
here:

https://github.com/andrewbutterfield/CS4098-2017/blob/84a0cbfb7931071f5ef8c2ed516f3289e1f6df83/DocPML.pdf

Our implementation differs in a few small regards:

`manual` and `executable` are allowed as variable names. These are supposedly
keywords in the language but there is some sample pml from PEOS where this
happens:

https://github.com/jnoll/peos/blob/master/compiler/models/edit-compile-test.pml#L9

The expression productions outlined in the spec have some odd characteristics.
For example, comparing a variable to a value is not allowed:

`variable == "string"`

But comparing an attribute of a variable is:

`variable.attribute == "string"`

Our implementation allows variables and values to be compared (seems more
sensisble).