initial attempt at DDI categorization
This adds the basic outline for categorizing DDIs. To do so it requires:

* an AST
* a list of DDIs `{drug_a, drug_b, ...}`
* a list of drugs `{uri, label}`

It iterates over the AST, mapping each drug to a list of the PML constructs it
is contained in (its `ancestors`).

For each DDI, it looks at the ancestors of the participating drugs, and gets
their intersection. It takes the common ancestor with the highest line number as
the `closest_common_ancestor`.

The type of this ancestor is used to categorize the DDI:

* `branch => parallel`
* `selection => alternative`
* `_ => sequential` # everything else

Caveats:

* does not yet handle drugs occurring more than once in a PML file
* does not yet properly handle multiple pml constructs on the same line:
  `task { task { ... } }`