adds basic handling of repeated drugs
Changes `Analysis.Ddis` so that it builds a separate ancestry for each
occurrence of a drug.

We probably want to change the parameters passed to `/api/ddis` to include the
line number of each drug. At the moment repeated drugs can lead to duplicate
DDIs because they are identified by the enclosing construct rather than the line
numbers of the participating drugs.