Adds time to Ast.ex
This change allows Ast.ex to deal with times in the PML AST. To
facilitate this, it changes `:requires` into a multi-line construct.
`:requires` is different to other multi-line constructs in that it has a
single child node instead of a list of children. To deal with this, I've
added another `do_unquote` clause that wraps the child node in a list.