changes how the parser handles numbers
This change was going to just add a test to ensure the parser properly
deals with non-integer time values. With the current lexer and parser
implementation, any error message containing a number would look like
nonsense - because yecc keeps the tokens as a charlist.

We expect a message like:

`syntax error before: 4.5`

But we'd actually get something like:

`syntax error before: [49,52,34]`

This changes how the lexer handles numbers to improve these error
messages. The lexer turns the character list representing a number into
an actual number (int or float). This allows the parser to print a nice
message. For simplicity, we've kept all expressions in the AST as
strings, so I've also changed the parser to turn the numbers back into
charlists when it's building the AST.